% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AcceptAffin.R
\name{AcceptAffin}
\alias{AcceptAffin}
\title{Calculates the "Acceptability Function" used in defining Blaker's (2000) Acceptability Interval and computing the latter in the function AcceptAffCI().}
\usage{
AcceptAffin(x, marg, alph)
}
\arguments{
\item{x}{integer co-occurrence count that should properly fall within the closed interval max(0,mA+mB-N), min(mA,mB)}

\item{marg}{a 3-entry integer vector (mA,mB,N) consisting of the first row and column totals and the table total for a 2x2 contingency table}

\item{alp}{a vector of (one or more) real-valued "alpha" values, where alpha  is the log-odds parameter in the Extended Hypergeometric distribution}
}
\value{
This function returns the "Acceptability Function" that is later used by another function AcceptAffCI() to compute "Acceptability Interval".
}
\description{
This function calculates the "Acceptability Function" of Blaker (2000, Thm.1, p.785) for the log-odds parameter alpha in the Extended Hypergeometric distribution.
}
\details{
This function calculates the "Acceptability Function" of Blaker (2000, Thm.1, p.785) for the log-odds parameter alpha
in the Extended Hypergeometric distribution, a function from which the "Acceptability Interval" is calculated by another CooccurrenceAffinity package function AcceptAffCI().
}
\references{
Blaker, H. (2000), “Confidence curves and improved exact confidence intervals for discrete distributions", Canadian Journal of Statistics 28, 783-798.
}
\author{
Eric Slud
}
