% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlphInts.R
\name{AlphInts}
\alias{AlphInts}
\title{Median interval, four confidence intervals, null expectation of cooccurrence count, p-value}
\usage{
AlphInts(x, marg, scal = log(2 * marg[3]^2), lev = 0.95)
}
\arguments{
\item{x}{integer co-occurrence counts that should properly fall within the closed interval  \link{max(0,mA+mB-N), min(mA,mB)}}

\item{marg}{a 3-entry integer vector containing (mA,mB,N)}

\item{scal}{an integer parameter (default 10) that should fall somewhere between 2 and 10}

\item{lev}{a confidence level, generally somewhere from 0.8 to 0.95  (default 0.95)}
}
\value{
This function returns various types of intervals including confidence intervals
}
\description{
This function calculates
(i) MedianIntrvl, the interval of alpha values for which the co-occurrence count is a median,
(ii) four Confidence Intervals, two using EHypQuInt(), one using EHypMidP(), and one using AcceptAffCI(),
(iii) the Expected Co-occurrence count under the Null distibution, and
(iv) the p-value for the observed co-occurrence count.
}
\details{
This function calculates five intervals, three of them using EHypQuInt, one using EHypMidP,
and one using AcceptAffCI. First ("Int1") is the interval of alpha values compatible with x as median for the
Extended Hypergeometric distribution (with fixed margins and alpha);
second ("Int2") an "exact" conservative test-based confidence intrval (in a sense analogous to the Clopper-Pearson (1934)
confidence interval for unknown binomial proportion) for alpha based on data (x,mA,mB,N);
third the Acceptability Confidence Interval ("Int3") of Blaker (2000)
which is a better confidence interval than the CP-type interval "Int2" in the sense of being contained within "Int2"
but still provably conservative.
The fourth confidence interval ("Int4") is the one given in formula (2) above of the Introduction to this documentation,
with endpoints obtained as the midpoints of quantile intervals respectively to the (1+lev)/2 and (1-lev)/2
quantiles of the Extended Hypergeometric distribution; and the fifth ("Int5")
which behaves very similarly to "Int4" is defined by the midP approach analogous to the midP confidence interval
for binomial proportions (Agresti 2013, p.605), and is calculated from EHypMidP.

The first of these intervals provides quantification of the underlying discreteness of the
Extended Hypergeometric and the impact of that discreteness on the estimation of alpha:
Int1 is an interval that will contain the MLE alpha-hat, and the mid-point of that interval is another reasonable estimator of alpha from the data.
The recommended (slightly conservative) confidence interval is Int3,
while the very similar intervals Int4 and Int5 have coverage typically closer than
Int2 or Int3 to the nominal level f coverage, at the cost of occasionally under-covering by as much as 0.04 or 0.05 for confidence levels 0.90 or 0.95.
}
\references{
to be added
}
\author{
Eric Slud
}
