% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bisect.R
\name{Bisect}
\alias{Bisect}
\title{Bisect}
\usage{
Bisect(ffnc, intrv, tol = 1e-08)
}
\arguments{
\item{x}{nteger co-occurrence counts that should properly fall within the closed interval  \link{max(0,mA+mB-N), min(mA,mB)}}

\item{marg}{a 3-entry integer vector containing (mA,mB,N)}

\item{scal}{an integer parameter (default 10) that should fall somewhere between 2 and 10}

\item{lev}{a confidence level, generally somewhere from 0.8 to 0.95  (default 0.95)}
}
\value{
This function returns
}
\description{
This function bisects...
}
\details{
This function bisects...
}
\references{
to be added
}
\author{
Eric Slud
}
