% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EHypMidP.R
\name{EHypMidP}
\alias{EHypMidP}
\title{Largest interval of alpha values-2}
\usage{
EHypMidP(marg, x, lev)
}
\arguments{
\item{marg}{a 3-entry integer vector containing (mA,mB,N)}

\item{x}{nteger co-occurrence counts that should properly fall within the closed interval  \link{max(0,mA+mB-N), min(mA,mB)}}

\item{lev}{a confidence level, generally somewhere from 0.8 to 0.95  (default 0.95)}
}
\value{
This function returns the largest interval of alpha values
}
\description{
This function does the analogous calculation to that of EHypQuInt,
but with the Extended Hypergeometric distribution function  F(x) = F(x,mA,mB,N, exp(alpha)) replaced by  (F(x) + F(x-1))/2.
}
\details{
This function does the analogous calculation to that of EHypQuInt,
but with the Extended Hypergeometric distribution function  F(x) = F(x,mA,mB,N, exp(alpha)) replaced by  (F(x) + F(x-1))/2.
}
\references{
to be added
}
\author{
Eric Slud
}
