% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EHypQuInt.R
\name{EHypQuInt}
\alias{EHypQuInt}
\title{Largest interval of alpha values}
\usage{
EHypQuInt(marg, x, q, scal = log(2 * marg[3]^2))
}
\arguments{
\item{marg}{a 3-entry integer vector containing (mA,mB,N)}

\item{x}{nteger co-occurrence counts that should properly fall within the closed interval  \link{max(0,mA+mB-N), min(mA,mB)}}

\item{q}{a quantile falling strictly between 0 and 1}

\item{scal}{an integer parameter (default 10) that should fall somewhere between 2 and 10}
}
\value{
This function returns the largest interval of alpha values
}
\description{
This function outputs the largest interval of alpha values within which the Extended Hypergeometric distribution function
at x is >= q  and the complementary distribution function 1 - F(x-) is >= 1-q.
}
\details{
This function outputs the largest interval of alpha values within which the Extended Hypergeometric distribution
function at x is >= q  and the complementary distribution function 1 - F(x-) is >= 1-q.
This is the set of alpha values "compatible" with x being a q'th quantile for the Extended Hypergeometric.
}
\references{
to be added
}
\author{
Eric Slud
}
