% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxX.Int.R
\name{MaxX.Int}
\alias{MaxX.Int}
\title{MaxX.Int}
\usage{
MaxX.Int(marg, scal = log(2 * marg[3]^2), lev = 0.95)
}
\arguments{
\item{marg}{a 3-entry integer vector (mA,mB,N) consisting of the first row and column totals and the table total for a 2x2 contingency table}

\item{scal}{an integer parameter (default 2*N^2, capped at 10 within the function) that should be 2 or greater}

\item{lev}{a confidence level, generally somewhere from 0.8 to 0.95  (default 0.95)}
}
\value{

}
\description{
Helper function
}
\details{
This is a helper function.
}
\references{

}
\author{
Eric Slud
}
