% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midP.EHyp.R
\name{midP.EHyp}
\alias{midP.EHyp}
\title{midP.EHyp computation}
\usage{
midP.EHyp(alp)
}
\arguments{
\item{alp}{"alpha" parameter, the log-odds parameter in the Extended Hypergeometric distribution}
}
\value{
helper function for midP CI computation with EHypMidP
}
\description{
Helper function
}
\details{
This is a helper function.

param x integer co-occurrence count that should properly fall within the closed interval [max(0,mA+mB-N), min(mA,mB)]
}
\author{
Eric Slud
}
