% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML.Alpha.R
\name{ML.Alpha}
\alias{ML.Alpha}
\title{Maximum likelihood estimate and intervals of alpha}
\usage{
ML.Alpha(x, marg, bound = T, scal = log(2 * marg[3]^2), lev = 0.95)
}
\arguments{
\item{x}{nteger co-occurrence counts that should properly fall within the closed interval  \link{max(0,mA+mB-N), min(mA,mB)}}

\item{marg}{a 3-entry integer vector containing (mA,mB,N)}

\item{scal}{an integer parameter (default 10) that should fall somewhere between 2 and 10}

\item{lev}{a confidence level, generally somewhere from 0.8 to 0.95  (default 0.95)}
}
\value{
This function returns maximum likelihood estimate and various confidence intervals of alpha
}
\description{
This function calculates the maximum likelihood estimate of alpha, which is the "affinity" index of co-occurrence
championed in our paper Mainali et al as an index of cooccurrence-based similarity, along with the intervals computed in AlphInts.
}
\details{
This function calculates the maximum likelihood estimate of alpha, which is the "affinity" index of co-occurrence
championed in our paper Mainali et al as an index of cooccurrence-based similarity, along with the intervals computed in AlphInts,
now called CI2 through CI5. The boolean "bound" parameter is an option to prevent the intervals containing alpha-estimates
to extend to plus or minus infinity, based on a Bayesian argument. The bound substituted for the Infinite endpoints is
provably larger than the largest value the MLE can take whenever x avoids the enpoints max(mA+mB-N,0) and min(mA,mB) of its logical range.
The recommended confidence interval for alpha is CI2.
}
\references{
to be added
}
\author{
Eric Slud
}
