% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.gg.R
\name{plot.gg}
\alias{plot.gg}
\title{Heatmap plot of affinity() output}
\usage{
\method{plot}{gg}(
  data,
  variable,
  legendlimit,
  col = NULL,
  show.value = NULL,
  value.digit = NULL,
  text.size = NULL,
  plot.margin = NULL,
  ...
)
}
\arguments{
\item{x}{integer co-occurrence count that should properly fall within the closed interval [max(0,mA+mB-N), min(mA,mB)]}

\item{marg}{a 3-entry integer vector (mA,mB,N) consisting of the first row and column totals and the table total for a 2x2 contingency table}

\item{bound}{a boolean parameter which when TRUE replaces the MLE of "+/-Infinity", applicable when x is respectively at the upper extreme min(mA,mB)
or the lower extreme max(mA+mB-N,0) of its possible range, by a finite value with absolute value upper-bounding the value of
MLEs attainable for values of x not equal to its extremes}

\item{scal}{an integer parameter (default 2*N^2, capped at 10 within the function) that should be 2 or greater}

\item{lev}{a confidence level, generally somewhere from 0.8 to 0.95  (default 0.95)}

\item{pvalType}{a character string telling what kind of p-value to calculate. ‘Blaker’ or “midP’.
If ‘pvalType=Blaker” (the default value), the p-value is calculated according to "Acceptability" function of Blaker (2000).
If ‘pvalType=midP’, the p-value is calculated using the same idea as the midP confidence interval.}
}
\value{
This function returns maximum likelihood estimate of alpha, the interval-endpoints of alpha values for which x is a median,
and four confidence intervals for alpha, described in detail under documentation for AlphInts().
In addition there are two output list-components for the null-distribution expected co-occurrence count and the p-value
for the test of the null hypothesis alpha=0, calculated as in AlphInts.
}
\description{
This function works on the output of affinity() and gives a heatmap plot for the numeric columns of $all dataframe except
the interval columns (median interval and confidence intervals) and confidence level (which is a constant for all pairs in one run of the code)
}
\details{
This function is really a ggplot behind the scene where I have taken care of many arguments for generating a useful plot.
It generates a plot for the lower triangle of an NxN square matrix where both row and columns carry the same set of entities,
such that all pairwise analyses are shown in the plot. The plots can be requested using the column names of $all of the main output of affinity().
of affinity.
}
\examples{
unlist(ML.Alpha(30,c(50,80,120), lev=0.9))
}
\references{
Fog, A. (2015), BiasedUrn: Biased Urn Model Distributions. R package version 1.07.

Harkness, W. (1965), “Properties of the extended hypergeometric distribution“, Annals of Mathematical Statistics, 36, 938-945.

Mainali, K., Slud, E., Singer, M. and Fagan, W. (2022), "A better index for analysis of co-occurrence and similarity", Science Advances, to appear.
}
\author{
Kumar Mainali
}
