% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AcceptAffCI.R
\name{AcceptAffCI}
\alias{AcceptAffCI}
\title{Acceptability Interval}
\usage{
AcceptAffCI(x, marg, lev, CPint)
}
\arguments{
\item{x}{nteger co-occurrence counts that should properly fall within the closed interval  \link{max(0,mA+mB-N), min(mA,mB)}}

\item{marg}{a 3-entry integer vector containing (mA,mB,N)}

\item{lev}{a confidence level, generally somewhere from 0.8 to 0.95  (default 0.95)}

\item{CPint}{the exact conservative ("Clopper-Pearson-type") interval calculated as $Int2 in the function AlphInts()}
}
\value{
This function returns the "Acceptability Interval"
}
\description{
This function calculates the "Acceptability Interval"
}
\details{
This function calculates the "Acceptability Interval" based on "Acceptability Function" computed by AcceptAffin().
}
\references{
to be added
}
\author{
Eric Slud
}
