% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AcceptAffin.R
\name{AcceptAffin}
\alias{AcceptAffin}
\title{Calculate the "Acceptability Function"}
\usage{
AcceptAffin(x, marg, alph, lev)
}
\arguments{
\item{x}{nteger co-occurrence counts that should properly fall within the closed interval  \link{max(0,mA+mB-N), min(mA,mB)}}

\item{marg}{a 3-entry integer vector containing (mA,mB,N)}

\item{lev}{a confidence level, generally somewhere from 0.8 to 0.95  (default 0.95)}

\item{alp}{a vector of (one or more) real "alpha" values}
}
\value{
This function returns the "Acceptability Function" that is later used by another function AcceptAffCI() to compute "Acceptability Interval"
}
\description{
This function calculates the "Acceptability Function" of Blaker (2000, Thm.1, p.785) for the Extended Hypergeometric distribution.
}
\details{
This function calculates the "Acceptability Function" of Blaker (2000, Thm.1, p.785) for the Extended Hypergeometric distribution,
a function from which the "Acceptability Interval: is calculated by another
Affinity package function  AcceptAffCI = function(x, marg, lev, CPint).
Here "CPint" is the exact conservative ("Clopper-Pearson-type") interval calculated as $Int2 in the function AlphInts below.
}
\references{
to be added
}
\author{
Eric Slud
}
