% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovrgPlot.R
\name{CovrgPlot}
\alias{CovrgPlot}
\title{Coverage probabilities of the confidence intervals}
\usage{
CovrgPlot(
  marg,
  scal = log(2 * marg[3]^2),
  lev = 0.95,
  alim = 4,
  minp = 0.7,
  intpts = 1,
  plotCI = 1:4
)
}
\arguments{
\item{marg}{a 3-entry integer vector containing (mA,mB,N)}

\item{scal}{an integer parameter (default 10) that should fall somewhere between 2 and 10}

\item{lev}{a confidence level, generally somewhere from 0.8 to 0.95  (default 0.95)}
}
\value{
This function returns maximum likelihood estimate and various confidence intervals of alpha
}
\description{
This function calculates and plots the coverage probabilities of the confidence intervals CI1 and CI2, which is useful in showing how conservative
and more nomimal recommended confidence inervals map against each other.
}
\details{
This function calculates and plots the coverage probabilities of the confidence intervals CI1 and CI2,
for fixed marginals "marg" = (mA,mB,N) and confidence level "lev" as a function of alpha.
The plots serve as a useful exhibit to show the range of always conservative (and sometimes quite conservative)
coverage probabilities for CI1, and the much closer-to-nominal coverage probabilities for the
recommended Confidence Interval CI2. The outputted vector "alphvec" of alpha values is the set of
all discontinuities in the slope of the coverag-probability function together with the corresponding exact (not simulated)
coverage probabilities of each of the two confidence intervals at the alphvec points in the two-column array  covPrb.
}
\references{
to be added
}
\author{
Eric Slud
}
