% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLikExtHyp.R
\name{logLikExtHyp}
\alias{logLikExtHyp}
\title{Repeat as stand-alone calculations of part of ML.Alpha()}
\usage{
logLikExtHyp(x, marg, alpha)
}
\arguments{
\item{x}{integer co-occurrence count that should properly fall within the closed interval [max(0,mA+mB-N), min(mA,mB)]}

\item{marg}{a 3-entry integer vector (mA,mB,N) consisting of the first row and column totals and the table total for a 2x2 contingency table}
}
\value{
to be added
}
\description{
This function calculates ... to be added
}
\details{
to be added
}
\references{
to be added
}
\author{
Eric Slud
}
