% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midP.EHyp.R
\name{midP.EHyp}
\alias{midP.EHyp}
\title{midP.EHyp}
\usage{
midP.EHyp(alp)
}
\arguments{
\item{x}{integer co-occurrence count that should properly fall within the closed interval [max(0,mA+mB-N), min(mA,mB)]}
}
\value{

}
\description{
Helper function
}
\details{
This is a helper function.
}
\references{

}
\author{
Eric Slud
}
